backend/requirements/base.txt
Django==2.1.7
djangorestframework==3.9.2
psycopg2-binary==2.7.7
django-rest-auth==0.9.3
djangorestframework-jwt==1.11.0
django-allauth==0.39.1